make S speeds 8bit integer from 0 .. 255. S validation and normalization should be offloaded to streaming applications for maximum Grbl efficiency

+++ grbl/gcode.h
@@ -165,7 +165,7 @@ typedef struct {
   float p;         // G10 or dwell parameters
   // float q;      // G82 peck drilling
   float r;         // Arc radius
-  float s;         // Spindle speed
+  uint8_t s;         // Spindle speed
   uint8_t t;       // Tool selection
   float xyz[3];    // X,Y,Z Translational axes
 } gc_values_t;
@@ -174,7 +174,7 @@ typedef struct {
 typedef struct {
   gc_modal_t modal;
   
-  float spindle_speed;          // RPM
+  uint8_t spindle_speed;          // RPM
   float feed_rate;              // Millimeters/min
   uint8_t tool;                 // Tracks tool number. NOT USED.
   int32_t line_number;          // Last line number sent
